change export name since directory is now an option
