+++ las_geoh5/export_files/__init__.py
@@ -1,0 +1,7 @@
+#  Copyright (c) 2024 Mira Geoscience Ltd.
+#
+#  This file is part of las-geoh5 project.
+#
+#  las-geoh5 is distributed under the terms and conditions of the MIT License
+#  (see LICENSE file at the root of this source code package).
+#
