+++ las_geoh5/uijson/write_uijson.py
@@ -30,8 +30,8 @@ def write_uijson(basepath: str | Path, mode: str = "import_files"):
     :return: Input file for the written data.
     """
 
-    if mode not in ["import_files", "import_directories", "export_directories"]:
-        msg = "Mode argument must be 'import_files', 'import_directories', or 'export_directories'."
+    if mode not in ["import_files", "import_directories", "export_files"]:
+        msg = "Mode argument must be 'import_files', 'import_directories', or 'export_files'."
         raise ValueError(msg)
 
     module = importlib.import_module(f"las_geoh5.{mode}.uijson")
@@ -53,10 +53,10 @@ def main(args):
     parser.add_argument(
         "mode",
         type=str,
-        choices={"import_files", "import_directories", "export_directories", "all"},
+        choices={"import_files", "import_directories", "export_files", "all"},
         help=(
             "Mode switching between 'import_files', 'import_directories',"
-            " 'export_directories', and 'all' behaviour."
+            " 'export_files', and 'all' behaviour."
         ),
     )
     args = parser.parse_args(args)
@@ -64,7 +64,7 @@ def main(args):
         for mode in [
             "import_files",
             "import_directories",
-            "export_directories",
+            "export_files",
         ]:  # pylint: disable=invalid-name
             write_uijson(args.path, mode)
     else:
